feat(components, protocol-designer): snackBar and toast in components library
closes AUTH-651
