+++ app/src/organisms/AdvancedSettings/ClearUnavailableRobots.tsx
@@ -16,13 +16,14 @@ import {
   SPACING_AUTO,
   SPACING,
   LegacyStyledText,
+  ERROR_TOAST,
+  SUCCESS_TOAST,
   TYPOGRAPHY,
   useConditionalConfirm,
   Modal,
 } from '@opentrons/components'
 
 import { TertiaryButton } from '../../atoms/buttons'
-import { ERROR_TOAST, SUCCESS_TOAST } from '../../atoms/Toast'
 import { useToaster } from '../../organisms/ToasterOven'
 import { getTopPortalEl } from '../../App/portal'
 import {
