+++ app/src/organisms/Alerts/AlertsModal.tsx
@@ -2,11 +2,11 @@ import * as React from 'react'
 import { useSelector, useDispatch } from 'react-redux'
 import { useTranslation } from 'react-i18next'
 import head from 'lodash/head'
+import { SUCCESS_TOAST, WARNING_TOAST } from '@opentrons/components'
 
 import * as AppAlerts from '../../redux/alerts'
 import { getHasJustUpdated, toggleConfigValue } from '../../redux/config'
 import { getAvailableShellUpdate } from '../../redux/shell'
-import { SUCCESS_TOAST, WARNING_TOAST } from '../../atoms/Toast'
 import { useToaster } from '../ToasterOven'
 import { UpdateAppModal } from '../UpdateAppModal'
 import { U2EDriverOutdatedAlert } from './U2EDriverOutdatedAlert'
