+++ app/src/organisms/Devices/RobotSettings/AdvancedTab/Troubleshooting.tsx
@@ -11,6 +11,8 @@ import {
   Flex,
   JUSTIFY_SPACE_BETWEEN,
   SPACING_AUTO,
+  ERROR_TOAST,
+  INFO_TOAST,
   SPACING,
   LegacyStyledText,
   TYPOGRAPHY,
@@ -18,7 +20,6 @@ import {
 import { useHost } from '@opentrons/react-api-client'
 
 import { TertiaryButton } from '../../../../atoms/buttons'
-import { ERROR_TOAST, INFO_TOAST } from '../../../../atoms/Toast'
 import { useToaster } from '../../../../organisms/ToasterOven'
 import { CONNECTABLE } from '../../../../redux/discovery'
 import { useRobot } from '../../hooks'
