+++ app/src/organisms/Devices/hooks/useDownloadRunLog.ts
@@ -2,7 +2,7 @@ import * as React from 'react'
 import { useTranslation } from 'react-i18next'
 import { getRun, getCommands, getProtocol } from '@opentrons/api-client'
 import { useHost } from '@opentrons/react-api-client'
-import { ERROR_TOAST, INFO_TOAST } from '../../../atoms/Toast'
+import { ERROR_TOAST, INFO_TOAST } from '@opentrons/components'
 import { useToaster } from '../../../organisms/ToasterOven'
 import { downloadFile } from '../utils'
 import type { IconProps } from '@opentrons/components'
