+++ components/src/atoms/Toast/__tests__/Toast.test.tsx
@@ -2,14 +2,11 @@ import * as React from 'react'
 import { describe, it, expect, vi, beforeEach } from 'vitest'
 import '@testing-library/jest-dom/vitest'
 import { act, fireEvent, screen } from '@testing-library/react'
-import { i18n } from '../../../i18n'
+import { renderWithProviders } from '../../../testing/utils'
 import { Toast, TOAST_ANIMATION_DURATION } from '..'
-import { renderWithProviders } from '../../../__testing-utils__'
 
 const render = (props: React.ComponentProps<typeof Toast>) => {
-  return renderWithProviders(<Toast {...props} displayType="desktop" />, {
-    i18nInstance: i18n,
-  })[0]
+  return renderWithProviders(<Toast {...props} displayType="desktop" />)[0]
 }
 
 describe('Toast', () => {
