+++ protocol-designer/src/ProtocolRoutes.tsx
@@ -7,6 +7,7 @@ import { Liquids } from './pages/Liquids'
 import { Designer } from './pages/Designer'
 import { CreateNewProtocolWizard } from './pages/CreateNewProtocolWizard'
 import { NavigationBar } from './NavigationBar'
+import { Kitchen } from './organisms'
 
 import type { RouteProps } from './types'
 
@@ -50,14 +51,16 @@ export function ProtocolRoutes(): JSX.Element {
   return (
     <>
       <NavigationBar routes={pdRoutes} />
-      <Box width="100%">
-        <Routes>
-          {allRoutes.map(({ Component, path }: RouteProps) => {
-            return <Route key={path} path={path} element={<Component />} />
-          })}
-          <Route path="*" element={<Navigate to={LANDING_ROUTE} />} />
-        </Routes>
-      </Box>
+      <Kitchen>
+        <Box width="100%">
+          <Routes>
+            {allRoutes.map(({ Component, path }: RouteProps) => {
+              return <Route key={path} path={path} element={<Component />} />
+            })}
+            <Route path="*" element={<Navigate to={LANDING_ROUTE} />} />
+          </Routes>
+        </Box>
+      </Kitchen>
     </>
   )
 }
