+++ protocol-designer/src/assets/localization/en/create_new_protocol.json
@@ -31,6 +31,7 @@
   "stagingArea_cutoutD3": "Staging area D3",
   "swap": "Swap pipettes",
   "tell_us": "Tell us about your protocol",
+  "up_to_3_tipracks": "Up to 3 tip rack types are allowed per pipette",
   "trashBin": "Trash Bin",
   "vol_label": "{{volume}} uL",
   "wasteChute": "Waste Chute",
