+++ protocol-designer/src/organisms/index.ts
@@ -1,1 +1,2 @@
+export * from './Kitchen'
 export * from './SlotInformation'
